split subcomponents of rigidtypes (coord, deriv, mass) in separate headers
