+++ Sofa/framework/DefaultType/src/sofa/defaulttype/RigidTypes.h
@@ -19,29 +19,23 @@
 *                                                                             *
 * Contact information: contact@sofa-framework.org                             *
 ******************************************************************************/
-#ifndef SOFA_DEFAULTTYPE_RIGIDTYPES_H
-#define SOFA_DEFAULTTYPE_RIGIDTYPES_H
+#pragma once
+#include <sofa/defaulttype/config.h>
 
-#include <sofa/defaulttype/fwd.h>
-#include <sofa/type/Vec.h>
+#include <sofa/defaulttype/RigidCoord.h>
+#include <sofa/defaulttype/RigidDeriv.h>
+#include <sofa/defaulttype/RigidMass.h>
 #include <sofa/defaulttype/MapMapSparseMatrix.h>
-#include <sofa/type/Mat.h>
+
+#include <sofa/type/Vec.h>
 #include <sofa/type/Quat.h>
 #include <sofa/type/vector.h>
 #include <sofa/helper/rmath.h>
 #include <sofa/helper/random.h>
-#include <iostream>
-#include <cstdlib>
 #include <cmath>
 
-#if !defined(NDEBUG)
-#include <sofa/helper/logging/Messaging.h>
-#endif
 
-namespace sofa
-{
-
-namespace defaulttype
+namespace sofa::defaulttype
 {
 
 //=============================================================================
@@ -50,778 +44,6 @@ namespace defaulttype
 
 /** Degrees of freedom of 3D rigid bodies. Orientations are modeled using quaternions.
  */
-template<typename real>
-class RigidDeriv<3, real>
-{
-public:
-    typedef real value_type;
-    typedef sofa::Size Size;
-    typedef real Real;
-    typedef type::Vec<3,Real> Pos;
-    typedef type::Vec<3,Real> Rot;
-    typedef type::Vec<3,Real> Vec3;
-    typedef type::Vec<6,Real> VecAll;
-    typedef type::Quat<Real> Quat;
-
-protected:
-    Vec3 vCenter;
-    Vec3 vOrientation;
-
-public:
-    friend class RigidCoord<3,real>;
-
-    RigidDeriv()
-    {
-        clear();
-    }
-
-    explicit RigidDeriv(type::NoInit)
-        : vCenter(type::NOINIT)
-        , vOrientation(type::NOINIT)
-    {
-
-    }
-
-    RigidDeriv(const Vec3 &velCenter, const Vec3 &velOrient)
-        : vCenter(velCenter), vOrientation(velOrient)
-    {}
-
-    template<typename real2>
-    RigidDeriv(const RigidDeriv<3,real2>& c)
-        : vCenter(c.getVCenter()), vOrientation(c.getVOrientation())
-    {}
-
-    template<typename real2>
-    RigidDeriv(const type::Vec<6,real2> &v)
-        : vCenter(type::Vec<3,real2>(v.data())), vOrientation(type::Vec<3,real2>(v.data()+3))
-    {}
-
-    template<typename real2>
-    RigidDeriv(const real2* ptr)
-        :vCenter(ptr),vOrientation(ptr+3)
-    {
-    }
-
-    void clear()
-    {
-        vCenter.clear();
-        vOrientation.clear();
-    }
-
-    template<typename real2>
-    void operator=(const RigidDeriv<3,real2>& c)
-    {
-        vCenter = c.getVCenter();
-        vOrientation = c.getVOrientation();
-    }
-
-    template<typename real2>
-    void operator=(const type::Vec<3,real2>& v)
-    {
-        vCenter = v;
-    }
-
-    template<typename real2>
-    void operator=(const type::Vec<6,real2>& v)
-    {
-        vCenter = v;
-        vOrientation = type::Vec<3,real2>(v.data()+3);
-    }
-
-    void operator+=(const RigidDeriv& a)
-    {
-        vCenter += a.vCenter;
-        vOrientation += a.vOrientation;
-    }
-
-    void operator-=(const RigidDeriv& a)
-    {
-        vCenter -= a.vCenter;
-        vOrientation -= a.vOrientation;
-    }
-
-    RigidDeriv<3,real> operator+(const RigidDeriv<3,real>& a) const
-    {
-        RigidDeriv d;
-        d.vCenter = vCenter + a.vCenter;
-        d.vOrientation = vOrientation + a.vOrientation;
-        return d;
-    }
-
-    template<typename real2>
-    void operator*=(real2 a)
-    {
-        vCenter *= a;
-        vOrientation *= a;
-    }
-
-    template<typename real2>
-    void operator/=(real2 a)
-    {
-        vCenter /= a;
-        vOrientation /= a;
-    }
-
-
-
-    RigidDeriv<3,real> operator-() const
-    {
-        return RigidDeriv(-vCenter, -vOrientation);
-    }
-
-    RigidDeriv<3,real> operator-(const RigidDeriv<3,real>& a) const
-    {
-        return RigidDeriv<3,real>(this->vCenter - a.vCenter, this->vOrientation-a.vOrientation);
-    }
-
-
-    /// dot product, mostly used to compute residuals as sqrt(x*x)
-    Real operator*(const RigidDeriv<3,real>& a) const
-    {
-        return vCenter[0]*a.vCenter[0]+vCenter[1]*a.vCenter[1]+vCenter[2]*a.vCenter[2]
-                +vOrientation[0]*a.vOrientation[0]+vOrientation[1]*a.vOrientation[1]
-                +vOrientation[2]*a.vOrientation[2];
-    }
-
-
-    /// Euclidean norm
-    real norm() const
-    {
-        return helper::rsqrt( vCenter*vCenter + vOrientation*vOrientation);
-    }
-
-
-    Vec3& getVCenter() { return vCenter; }
-    Vec3& getVOrientation() { return vOrientation; }
-    const Vec3& getVCenter() const { return vCenter; }
-    const Vec3& getVOrientation() const { return vOrientation; }
-
-    Vec3& getLinear() { return vCenter; }
-    const Vec3& getLinear() const { return vCenter; }
-    Vec3& getAngular() { return vOrientation; }
-    const Vec3& getAngular() const { return vOrientation; }
-
-    VecAll getVAll() const
-    {
-        return VecAll(vCenter, vOrientation);
-    }
-
-    /// Velocity at point p, where p is the offset from the origin of the frame, given in the same coordinate system as the velocity of the origin.
-    Vec3 velocityAtRotatedPoint(const Vec3& p) const
-    {
-        return vCenter - cross(p, vOrientation);
-    }
-
-    /// write to an output stream
-    inline friend std::ostream& operator << ( std::ostream& out, const RigidDeriv<3,real>& v )
-    {
-        out<<v.vCenter<<" "<<v.vOrientation;
-        return out;
-    }
-    /// read from an input stream
-    inline friend std::istream& operator >> ( std::istream& in, RigidDeriv<3,real>& v )
-    {
-        in>>v.vCenter>>v.vOrientation;
-        return in;
-    }
-
-    /// Compile-time constant specifying the number of scalars within this vector (equivalent to the size() method)
-    static constexpr sofa::Size total_size = 6;
-
-    /// Compile-time constant specifying the number of dimensions of space (NOT equivalent to total_size for rigids)
-    static constexpr sofa::Size spatial_dimensions = 3;
-
-    real* ptr() { return vCenter.ptr(); }
-    const real* ptr() const { return vCenter.ptr(); }
-
-    static Size size() {return 6;}
-
-    /// Access to i-th element.
-    real& operator[](Size i)
-    {
-        if (i<3)
-            return this->vCenter(i);
-        else
-            return this->vOrientation(i-3);
-    }
-
-    /// Const access to i-th element.
-    const real& operator[](Size i) const
-    {
-        if (i<3)
-            return this->vCenter(i);
-        else
-            return this->vOrientation(i-3);
-    }
-
-    /// @name Tests operators
-    /// @{
-
-    bool operator==(const RigidDeriv<3,real>& b) const
-    {
-        return vCenter == b.vCenter && vOrientation == b.vOrientation;
-    }
-
-    bool operator!=(const RigidDeriv<3,real>& b) const
-    {
-        return vCenter != b.vCenter || vOrientation != b.vOrientation;
-    }
-
-    /// @}
-
-};
-
-template<typename real, typename real2>
-inline  RigidDeriv<3,real> operator*(RigidDeriv<3, real> r, real2 a)
-{
-    r*=a;
-    return r;
-}
-
-template<typename real, typename real2>
-inline RigidDeriv<3,real> operator/(RigidDeriv<3, real> r,real2 a)
-{
-    r/=a;
-    return r;
-}
-
-template<sofa::Size N,typename T>
-typename RigidDeriv<N,T>::Pos& getLinear(RigidDeriv<N,T>& v)
-{
-    return v.getLinear();
-}
-
-template<sofa::Size N, typename T>
-const typename RigidDeriv<N,T>::Pos& getLinear(const RigidDeriv<N,T>& v)
-{
-    return v.getLinear();
-}
-
-template<sofa::Size N, typename T>
-typename RigidDeriv<N,T>::Rot& getAngular(RigidDeriv<N,T>& v)
-{
-    return v.getAngular();
-}
-
-template<sofa::Size N, typename T>
-const typename RigidDeriv<N,T>::Rot& getAngular(const RigidDeriv<N,T>& v)
-{
-    return v.getAngular();
-}
-
-template<sofa::Size N,typename T>
-typename RigidDeriv<N,T>::Pos& getVCenter(RigidDeriv<N,T>& v)
-{
-    return v.getLinear();
-}
-
-template<sofa::Size N, typename T>
-const typename RigidDeriv<N,T>::Pos& getVCenter(const RigidDeriv<N,T>& v)
-{
-    return v.getLinear();
-}
-
-template<sofa::Size N, typename T>
-typename RigidDeriv<N,T>::Rot& getVOrientation(RigidDeriv<N,T>& v)
-{
-    return v.getAngular();
-}
-
-template<sofa::Size N, typename T>
-const typename RigidDeriv<N,T>::Rot& getVOrientation(const RigidDeriv<N,T>& v)
-{
-    return v.getAngular();
-}
-
-/// Velocity at point p, where p is the offset from the origin of the frame, given in the same coordinate system as the velocity of the origin.
-template<typename T, typename R>
-type::Vec<3,T> velocityAtRotatedPoint(const RigidDeriv<3,R>& v, const type::Vec<3,T>& p)
-{
-    return getLinear(v) + cross( getAngular(v),p );
-}
-
-template<typename real>
-class RigidCoord<3,real>
-{
-public:
-    typedef real value_type;
-    typedef sofa::Size Size;
-    typedef real Real;
-    typedef type::Vec<3,Real> Pos;
-    typedef type::Quat<Real> Rot;
-    typedef type::Vec<3,Real> Vec3;
-    typedef type::Quat<Real> Quat;
-    typedef RigidDeriv<3,Real> Deriv;
-    typedef type::Mat<4,4,Real> HomogeneousMat;
-    typedef type::Vec<4,Real> HomogeneousVec;
-
-protected:
-    Vec3 center;
-    Quat orientation;
-public:
-    RigidCoord (const Vec3 &posCenter, const Quat &orient)
-        : center(posCenter), orientation(orient) {}
-    RigidCoord () { clear(); }
-
-    template<typename real2>
-    RigidCoord(const RigidCoord<3,real2>& c)
-        : center(c.getCenter()), orientation(c.getOrientation())
-    {
-    }
-
-
-    void clear() { center.clear(); orientation.clear(); }
-
-    /**
-     * @brief Random rigid transform composed of 3 random translations and 3 random Euler angles
-     * @param a Range of each random value: (-a,+a)
-     * @return random rigid transform
-     */
-    static RigidCoord rand(SReal a)
-    {
-        RigidCoord t;
-        t.center = Pos(  SReal(helper::drand(a)), SReal(helper::drand(a)), SReal(helper::drand(a)) );
-        t.orientation = Quat::fromEuler( SReal(helper::drand(a)),SReal(helper::drand(a)), SReal(helper::drand(a)));
-        return t;
-    }
-
-    template<typename real2>
-    void operator=(const RigidCoord<3,real2>& c)
-    {
-        center = c.getCenter();
-        orientation = c.getOrientation();
-    }
-
-    void operator =(const Vec3& p)
-    {
-        center = p;
-    }
-
-    //template<typename real2>
-    //void operator =(const RigidCoord<3,real2>& c)
-    //{
-    //    center = c.getCenter();
-    //    orientation = c.getOrientation();
-    //}
-
-    void operator +=(const Deriv& dg) {
-        // R3 x SO(3) exponential integration
-        center += dg.getVCenter();
-
-        const Vec3 omega = dg.getVOrientation() / 2;
-        const real theta = omega.norm();
-
-        static const real epsilon = std::numeric_limits<real>::epsilon();
-
-        if( theta < epsilon ) {
-            // fallback to gnomonic projection
-            Quat exp(omega[0], omega[1], omega[2], 1);
-            exp.normalize();
-            orientation = exp * orientation;
-        } else {
-            // expontential
-            const real sinc = std::sin(theta) / theta;
-            const Quat exp(sinc * omega[0],
-                           sinc * omega[1],
-                           sinc * omega[2],
-                           std::cos(theta));
-            orientation = exp * orientation;
-        }
-
-    }
-
-    RigidCoord<3,real> operator+(const Deriv& dg) const {
-        RigidCoord c = *this;
-        c += dg;
-        return c;
-    }
-
-
-    RigidCoord<3,real> operator-(const RigidCoord<3,real>& a) const
-    {
-        return RigidCoord<3,real>(this->center - a.getCenter(), a.orientation.inverse() * this->orientation);
-    }
-
-    RigidCoord<3,real> operator+(const RigidCoord<3,real>& a) const
-    {
-        return RigidCoord<3,real>(this->center + a.getCenter(), a.orientation * this->orientation);
-    }
-
-    RigidCoord<3,real> operator-() const
-    {
-        return RigidCoord<3,real>( -this->center, this->orientation.inverse() );
-    }
-
-    void operator +=(const RigidCoord<3,real>& a)
-    {
-        center += a.getCenter();
-        orientation *= a.getOrientation();
-    }
-
-    template<typename real2>
-    void operator*=(real2 a)
-    {
-        center *= a;
-    }
-
-    template<typename real2>
-    void operator/=(real2 a)
-    {
-        center /= a;
-    }
-
-    template<typename real2>
-    RigidCoord<3,real> operator*(real2 a) const
-    {
-        RigidCoord r = *this;
-        r*=a;
-        return r;
-    }
-
-    template<typename real2>
-    RigidCoord<3,real> operator/(real2 a) const
-    {
-        RigidCoord r = *this;
-        r/=a;
-        return r;
-    }
-
-
-
-    /// dot product, mostly used to compute residuals as sqrt(x*x)
-    Real operator*(const RigidCoord<3,real>& a) const
-    {
-        return center[0]*a.center[0]+center[1]*a.center[1]+center[2]*a.center[2]
-                +orientation[0]*a.orientation[0]+orientation[1]*a.orientation[1]
-                +orientation[2]*a.orientation[2]+orientation[3]*a.orientation[3];
-    }
-
-    /** Squared norm. For the rotation we use the xyz components of the quaternion.
-    Note that this is not equivalent to the angle, so a 2d rotation and the equivalent 3d rotation have different norms.
-      */
-    real norm2() const
-    {
-        return center*center
-                + orientation[0]*orientation[0]
-                + orientation[1]*orientation[1]
-                + orientation[2]*orientation[2]; // xyzw quaternion has null x,y,z if rotation is null
-    }
-
-    /// Euclidean norm
-    real norm() const
-    {
-        return helper::rsqrt(norm2());
-    }
-
-
-    Vec3& getCenter () { return center; }
-    Quat& getOrientation () { return orientation; }
-    const Vec3& getCenter () const { return center; }
-    const Quat& getOrientation () const { return orientation; }
-
-    static RigidCoord<3,real> identity()
-    {
-        RigidCoord c;
-        return c;
-    }
-
-    Vec3 rotate(const Vec3& v) const
-    {
-        return orientation.rotate(v);
-    }
-    Vec3 inverseRotate(const Vec3& v) const
-    {
-        return orientation.inverseRotate(v);
-    }
-
-    Vec3 translate(const Vec3& v) const
-    {
-        return v + center;
-    }
-
-    /// Apply a transformation with respect to itself
-    void multRight( const RigidCoord<3,real>& c )
-    {
-        center += orientation.rotate(c.getCenter());
-        orientation = orientation * c.getOrientation();
-    }
-
-    /// compute the product with another frame on the right
-    RigidCoord<3,real> mult( const RigidCoord<3,real>& c ) const
-    {
-        RigidCoord r;
-        r.center = center + orientation.rotate( c.center );
-        r.orientation = orientation * c.getOrientation();
-        return r;
-    }
-
-    /// Set from the given matrix
-    template<class Mat>
-    void fromMatrix(const Mat& m)
-    {
-        center[0] = m[0][3];
-        center[1] = m[1][3];
-        center[2] = m[2][3];
-        sofa::type::Mat<3,3,Real> rot; rot = m;
-        orientation.fromMatrix(rot);
-    }
-
-    /// Write to the given matrix
-    template<class Mat>
-    void toMatrix( Mat& m) const
-    {
-        m.identity();
-        orientation.toMatrix(m);
-        m[0][3] = (typename Mat::Real)center[0];
-        m[1][3] = (typename Mat::Real)center[1];
-        m[2][3] = (typename Mat::Real)center[2];
-    }
-
-    void toHomogeneousMatrix( HomogeneousMat& m) const
-    {
-        m.identity();
-        orientation.toHomogeneousMatrix(m);
-        m[0][3] = center[0];
-        m[1][3] = center[1];
-        m[2][3] = center[2];
-    }
-
-    /// create a homogeneous vector from a 3d vector
-    template <typename V>
-    static HomogeneousVec toHomogeneous( V v, real r=1. ){
-        return HomogeneousVec( v[0], v[1], v[2], r );
-    }
-    /// create a 3d vector from a homogeneous vector
-    template <typename V>
-    static Vec3 fromHomogeneous( V v ){
-        return Vec3( v[0], v[1], v[2] );
-    }
-
-
-    template<class Mat>
-    void writeRotationMatrix( Mat& m) const
-    {
-        orientation.toMatrix(m);
-    }
-
-    /// Write the OpenGL transformation matrix
-    void writeOpenGlMatrix( float m[16] ) const
-    {
-        orientation.writeOpenGlMatrix(m);
-        m[12] = (float)center[0];
-        m[13] = (float)center[1];
-        m[14] = (float)center[2];
-    }
-
-    /// Apply the transform to a point, i.e. project a point from the child frame to the parent frame (translation and rotation)
-    Vec3 projectPoint( const Vec3& v ) const
-    {
-        return orientation.rotate(v)+center;
-    }
-
-    /// Apply the transform to a vector, i.e. project a vector from the child frame to the parent frame (rotation only, no translation added)
-    Vec3 projectVector( const Vec3& v ) const
-    {
-        return orientation.rotate(v);
-    }
-
-    /// Apply the inverse transform to a point, i.e. project a point from the parent frame to the child frame (translation and rotation)
-    Vec3 unprojectPoint( const Vec3& v ) const
-    {
-        return orientation.inverseRotate(v-center);
-    }
-
-    ///  Apply the inverse transform to a vector, i.e. project a vector from the parent frame to the child frame (rotation only, no translation)
-    Vec3 unprojectVector( const Vec3& v ) const
-    {
-        return orientation.inverseRotate(v);
-    }
-
-    /// obsolete. Use projectPoint.
-    Vec3 pointToParent( const Vec3& v ) const { return projectPoint(v); }
-    /// obsolete. Use unprojectPoint.
-    Vec3 pointToChild( const Vec3& v ) const { return unprojectPoint(v); }
-
-
-    /// write to an output stream
-    inline friend std::ostream& operator << ( std::ostream& out, const RigidCoord<3,real>& v )
-    {
-        out<<v.center<<" "<<v.orientation;
-        return out;
-    }
-    /// read from an input stream
-    inline friend std::istream& operator >> ( std::istream& in, RigidCoord<3,real>& v )
-    {
-        in>>v.center>>v.orientation;
-#if !defined(NDEBUG)
-        if (!v.orientation.isNormalized())
-        {
-            std::stringstream text;
-            text << "Rigid Object with invalid quaternion (non-unitary norm)! Normalising quaternion value... " << msgendl;
-            text << "Previous value was: " << v.orientation << msgendl ;
-            v.orientation.normalize();
-            text << "New value is: " << v.orientation;
-            msg_warning("Rigid") << text.str();
-        }
-#endif // NDEBUG
-        return in;
-    }
-    static Size max_size()
-    {
-        return 3;
-    }
-
-    /// Compile-time constant specifying the number of scalars within this vector (equivalent to the size() method)
-    static constexpr sofa::Size total_size = 7;
-    /// Compile-time constant specifying the number of dimensions of space (NOT equivalent to total_size for rigids)
-    static constexpr sofa::Size spatial_dimensions = 3;
-
-    real* ptr() { return center.ptr(); }
-    const real* ptr() const { return center.ptr(); }
-
-    static Size size() {return 7;}
-
-    /// Access to i-th element.
-    real& operator[](Size i)
-    {
-        if (i<3)
-            return this->center(i);
-        else
-            return this->orientation[i-3];
-    }
-
-    /// Const access to i-th element.
-    const real& operator[](Size i) const
-    {
-        if (i<3)
-            return this->center(i);
-        else
-            return this->orientation[i-3];
-    }
-
-    /// @name Tests operators
-    /// @{
-
-    bool operator==(const RigidCoord<3,real>& b) const
-    {
-        return center == b.center && orientation == b.orientation;
-    }
-
-    bool operator!=(const RigidCoord<3,real>& b) const
-    {
-        return center != b.center || orientation != b.orientation;
-    }
-
-    /// @}
-
-};
-
-template<typename real>
-class RigidMass<3, real>
-{
-public:
-    typedef real value_type;
-    typedef real Real;
-    typedef type::Mat<3,3,Real> Mat3x3;
-    Real mass,volume;
-    Mat3x3 inertiaMatrix;	      // Inertia matrix of the object
-    Mat3x3 inertiaMassMatrix;    // Inertia matrix of the object * mass of the object
-    Mat3x3 invInertiaMatrix;	  // inverse of inertiaMatrix
-    Mat3x3 invInertiaMassMatrix; // inverse of inertiaMassMatrix
-    RigidMass(Real m=1)
-    {
-        mass = m;
-        volume = 1;
-        inertiaMatrix.identity();
-        recalc();
-    }
-    void operator=(Real m)
-    {
-        mass = m;
-        recalc();
-    }
-    void operator+=(Real m)
-    {
-        mass += m;
-        recalc();
-    }
-    void operator-=(Real m)
-    {
-        mass -= m;
-        recalc();
-    }
-    // operator to cast to const Real
-    operator const Real() const
-    {
-        return mass;
-    }
-    void recalc()
-    {
-        inertiaMassMatrix = inertiaMatrix * mass;
-        const bool canInvert1 = invInertiaMatrix.invert(inertiaMatrix);
-        const bool canInvert2 = invInertiaMassMatrix.invert(inertiaMassMatrix);
-        assert(canInvert1);
-        assert(canInvert2);
-        SOFA_UNUSED(canInvert1);
-        SOFA_UNUSED(canInvert2);
-    }
-
-    inline friend std::ostream& operator << (std::ostream& out, const RigidMass<3, real>& m )
-    {
-        out<<m.mass;
-        out<<" "<<m.volume;
-        out<<" "<<m.inertiaMatrix;
-        return out;
-    }
-    inline friend std::istream& operator >> (std::istream& in, RigidMass<3, real>& m )
-    {
-        in>>m.mass;
-        in>>m.volume;
-        in>>m.inertiaMatrix;
-        return in;
-    }
-    void operator *=(Real fact)
-    {
-        mass *= fact;
-        inertiaMassMatrix *= fact;
-        invInertiaMassMatrix /= fact;
-    }
-    void operator /=(Real fact)
-    {
-        mass /= fact;
-        inertiaMassMatrix /= fact;
-        invInertiaMassMatrix *= fact;
-    }
-};
-
-template<typename real>
-inline RigidDeriv<3,real> operator*(const RigidDeriv<3,real>& d, const RigidMass<3,real>& m)
-{
-    RigidDeriv<3,real> res;
-    getVCenter(res) = getVCenter(d) * m.mass;
-    getVOrientation(res) = m.inertiaMassMatrix * getVOrientation(d);
-    return res;
-}
-
-template<typename real>
-inline RigidDeriv<3,real> operator*(const RigidMass<3,real>& m, const RigidDeriv<3,real>& d)
-{
-    RigidDeriv<3,real> res;
-    getVCenter(res) = getVCenter(d) * m.mass;
-    getVOrientation(res) = m.inertiaMassMatrix * getVOrientation(d);
-    return res;
-}
-
-template<typename real>
-inline RigidDeriv<3,real> operator/(const RigidDeriv<3,real>& d, const RigidMass<3, real>& m)
-{
-    RigidDeriv<3,real> res;
-    getVCenter(res) = getVCenter(d) / m.mass;
-    getVOrientation(res) = m.invInertiaMassMatrix * getVOrientation(d);
-    return res;
-}
-
-
 template<typename real>
 class StdRigidTypes<3, real>
 {
@@ -832,6 +54,7 @@ public:
     typedef typename Coord::Vec3 Vec3;
     typedef typename Coord::Quat Quat;
     typedef type::Vec<3,Real> AngularVector;
+    typedef MapMapSparseMatrix<Deriv> MatrixDeriv;
 
     static constexpr sofa::Size spatial_dimensions = Coord::spatial_dimensions;
     static constexpr sofa::Size coord_total_size = Coord::total_size;
@@ -850,8 +73,6 @@ public:
     static void setDPos(Deriv& d, const DPos& v) { getVCenter(d) = v; }
     static const DRot& getDRot(const Deriv& d) { return getVOrientation(d); }
     static void setDRot(Deriv& d, const DRot& v) { getVOrientation(d) = v; }
-
-    typedef MapMapSparseMatrix<Deriv> MatrixDeriv;
 
     typedef type::vector<Coord> VecCoord;
     typedef type::vector<Deriv> VecDeriv;
@@ -1019,10 +240,7 @@ public:
 };
 
 typedef StdRigidTypes<3,double> Rigid3dTypes;
-typedef RigidMass<3,double> Rigid3dMass;
-
 typedef StdRigidTypes<3,float> Rigid3fTypes;
-typedef RigidMass<3,float> Rigid3fMass;
 
 /// We now use template aliases so we do not break backward compatibility.
 template<> constexpr const char* Rigid3dTypes::Name() { return "Rigid3d"; }
@@ -1030,687 +248,10 @@ template<> constexpr const char* Rigid3fTypes::Name() { return "Rigid3f"; }
 
 typedef StdRigidTypes<3,SReal> Rigid3Types;  ///< un-defined precision type
 typedef StdRigidTypes<3,SReal> RigidTypes;   ///< alias (beurk)
-typedef RigidMass<3,SReal>     Rigid3Mass;   ///< un-defined precision type
 
 //=============================================================================
 // 2D Rigids
 //=============================================================================
-
-template<typename real>
-class RigidDeriv<2,real>
-{
-public:
-    typedef real value_type;
-    typedef sofa::Size Size;
-    typedef real Real;
-    typedef type::Vec<2,Real> Pos;
-    typedef Real Rot;
-    typedef type::Vec<2,Real> Vec2;
-    typedef type::Vec<3,Real> VecAll;
-
-private:
-    Vec2 vCenter;
-    Real vOrientation;
-
-public:
-    friend class RigidCoord<2,real>;
-
-    explicit RigidDeriv(type::NoInit)
-        : vCenter(type::NOINIT)
-        , vOrientation(type::NOINIT)
-    {
-
-    }
-
-    RigidDeriv()
-    {
-        clear();
-    }
-
-    template<typename real2>
-    RigidDeriv(const type::Vec<2,real2> &velCenter, const real2 &velOrient)
-        : vCenter(velCenter), vOrientation((Real)velOrient)
-    {}
-
-    template<typename real2>
-    RigidDeriv(const type::Vec<3,real2> &v)
-        : vCenter(type::Vec<2,real2>(v.data())), vOrientation((Real)v[2])
-    {}
-
-    void clear()
-    {
-        vCenter.clear();
-        vOrientation=0;
-    }
-
-    template<typename real2>
-    void operator=(const RigidDeriv<2,real2>& c)
-    {
-        vCenter = c.getVCenter();
-        vOrientation = (Real)c.getVOrientation();
-    }
-
-    template<typename real2>
-    void operator=(const type::Vec<2,real2>& v)
-    {
-        vCenter = v;
-    }
-
-    template<typename real2>
-    void operator=(const type::Vec<3,real2>& v)
-    {
-        vCenter = v;
-        vOrientation = (Real)v[2];
-    }
-
-    void operator+=(const RigidDeriv<2,real>& a)
-    {
-        vCenter += a.vCenter;
-        vOrientation += a.vOrientation;
-    }
-
-    void operator-=(const RigidDeriv<2,real>& a)
-    {
-        vCenter -= a.vCenter;
-        vOrientation -= a.vOrientation;
-    }
-
-    RigidDeriv<2,real> operator+(const RigidDeriv<2,real>& a) const
-    {
-        RigidDeriv<2,real> d;
-        d.vCenter = vCenter + a.vCenter;
-        d.vOrientation = vOrientation + a.vOrientation;
-        return d;
-    }
-
-    RigidDeriv<2,real> operator-(const RigidDeriv<2,real>& a) const
-    {
-        RigidDeriv<2,real> d;
-        d.vCenter = vCenter - a.vCenter;
-        d.vOrientation = vOrientation - a.vOrientation;
-        return d;
-    }
-
-    template<typename real2>
-    void operator*=(real2 a)
-    {
-        vCenter *= a;
-        vOrientation *= (Real)a;
-    }
-
-    template<typename real2>
-    void operator/=(real2 a)
-    {
-        vCenter /= a;
-        vOrientation /= (Real)a;
-    }
-
-
-
-    RigidDeriv<2,real> operator-() const
-    {
-        return RigidDeriv<2,real>(-vCenter, -vOrientation);
-    }
-
-    /// dot product, mostly used to compute residuals as sqrt(x*x)
-    Real operator*(const RigidDeriv<2,real>& a) const
-    {
-        return vCenter[0]*a.vCenter[0]+vCenter[1]*a.vCenter[1]
-                +vOrientation*a.vOrientation;
-    }
-
-    /// Euclidean norm
-    Real norm() const
-    {
-        return helper::rsqrt( vCenter*vCenter + vOrientation*vOrientation);
-    }
-
-    Vec2& getVCenter() { return vCenter; }
-    Real& getVOrientation() { return vOrientation; }
-    const Vec2& getVCenter() const { return vCenter; }
-    const Real& getVOrientation() const { return vOrientation; }
-
-    Vec2& getLinear() { return vCenter; }
-    Real& getAngular() { return vOrientation; }
-    const Vec2& getLinear() const { return vCenter; }
-    const Real& getAngular() const { return vOrientation; }
-
-    VecAll getVAll() const
-    {
-        return VecAll(vCenter, vOrientation);
-    }
-
-    /// Velocity at point p, where p is the offset from the origin of the frame, given in the same coordinate system as the velocity of the origin.
-    Vec2 velocityAtRotatedPoint(const Vec2& p) const
-    {
-        return vCenter + Vec2(-p[1], p[0]) * vOrientation;
-    }
-
-    /// write to an output stream
-    inline friend std::ostream& operator << ( std::ostream& out, const RigidDeriv<2,real>& v )
-    {
-        out<<v.vCenter<<" "<<v.vOrientation;
-        return out;
-    }
-    /// read from an input stream
-    inline friend std::istream& operator >> ( std::istream& in, RigidDeriv<2,real>& v )
-    {
-        in>>v.vCenter>>v.vOrientation;
-        return in;
-    }
-
-    /// Compile-time constant specifying the number of scalars within this vector (equivalent to the size() method)
-    enum { total_size = 3 };
-    /// Compile-time constant specifying the number of dimensions of space (NOT equivalent to total_size for rigids)
-    enum { spatial_dimensions = 2 };
-
-    real* ptr() { return vCenter.ptr(); }
-    const real* ptr() const { return vCenter.ptr(); }
-
-    static Size size() {return 3;}
-
-    /// Access to i-th element.
-    real& operator[](Size i)
-    {
-        if (i<2)
-            return this->vCenter(i);
-        else
-            return this->vOrientation;
-    }
-
-    /// Const access to i-th element.
-    const real& operator[](Size i) const
-    {
-        if (i<2)
-            return this->vCenter(i);
-        else
-            return this->vOrientation;
-    }
-
-
-    /// @name Tests operators
-    /// @{
-
-    bool operator==(const RigidDeriv<2,real>& b) const
-    {
-        return vCenter == b.vCenter && vOrientation == b.vOrientation;
-    }
-
-    bool operator!=(const RigidDeriv<2,real>& b) const
-    {
-        return vCenter != b.vCenter || vOrientation != b.vOrientation;
-    }
-
-    /// @}
-};
-
-template<typename real, typename real2>
-inline RigidDeriv<2,real> operator*(RigidDeriv<2,real> r, real2 a)
-{
-    r *= a;
-    return r;
-}
-
-template<typename real, typename real2>
-inline RigidDeriv<2,real> operator/(RigidDeriv<2,real> r, real2 a)
-{
-    r /= a;
-    return r;
-}
-
-/// Velocity at point p, where p is the offset from the origin of the frame, given in the same coordinate system as the velocity of the origin.
-template<typename R, typename T>
-type::Vec<2,R> velocityAtRotatedPoint(const RigidDeriv<2,T>& v, const type::Vec<2,R>& p)
-{
-    return getVCenter(v) + type::Vec<2,R>(-p[1], p[0]) * getVOrientation(v);
-}
-
-
-
-template<typename real>
-class RigidCoord<2,real>
-{
-public:
-    typedef real value_type;
-    typedef sofa::Size Size;
-    typedef real Real;
-    typedef type::Vec<2,Real> Pos;
-    typedef Real Rot;
-    typedef type::Vec<2,Real> Vec2;
-    typedef type::Mat<3,3,Real> HomogeneousMat;
-    typedef type::Vec<3,Real> HomogeneousVec;
-private:
-    Vec2 center;
-    Real orientation;
-public:
-    RigidCoord (const Vec2 &posCenter, const Real &orient)
-        : center(posCenter), orientation(orient) {}
-    RigidCoord () { clear(); }
-
-    void clear() { center.clear(); orientation = 0; }
-
-    /**
-     * @brief Random rigid transform composed of 2 random translations and a random angle
-     * @param a Range of each random value: (-a,+a)
-     * @return random rigid transform
-     */
-    static RigidCoord rand(SReal a)
-    {
-        RigidCoord t;
-        t.center = Pos( SReal(helper::drand(a)), SReal(helper::drand(a)));
-        t.orientation = SReal(helper::drand(a));
-        return t;
-    }
-
-    void operator +=(const RigidDeriv<2,real>& a)
-    {
-        center += getVCenter(a);
-        orientation += getVOrientation(a);
-    }
-
-    RigidCoord<2,real> operator + (const RigidDeriv<2,real>& a) const
-    {
-        RigidCoord<2,real> c = *this;
-        c.center += getVCenter(a);
-        c.orientation += getVOrientation(a);
-        return c;
-    }
-
-    RigidCoord<2,real> operator -(const RigidCoord<2,real>& a) const
-    {
-        return RigidCoord<2,real>(this->center - a.getCenter(), this->orientation - a.orientation);
-    }
-
-    RigidCoord<2,real> operator +(const RigidCoord<2,real>& a) const
-    {
-        return RigidCoord<2,real>(this->center + a.getCenter(), this->orientation + a.orientation);
-    }
-
-
-    RigidCoord<2,real> operator-() const
-    {
-        return RigidCoord<2,real>( -this->center, this->orientation.inverse() );
-    }
-
-
-    void operator +=(const RigidCoord<2,real>& a)
-    {
-        center += a.getCenter();
-        orientation += a.getOrientation();
-    }
-
-    template<typename real2>
-    void operator*=(real2 a)
-    {
-        center *= a;
-        orientation *= (Real)a;
-    }
-
-    template<typename real2>
-    void operator/=(real2 a)
-    {
-        center /= a;
-        orientation /= (Real)a;
-    }
-
-    template<typename real2>
-    RigidCoord<2,real> operator*(real2 a) const
-    {
-        RigidCoord<2,real> r = *this;
-        r *= a;
-        return r;
-    }
-
-    template<typename real2>
-    RigidCoord<2,real> operator/(real2 a) const
-    {
-        RigidCoord<2,real> r = *this;
-        r /= a;
-        return r;
-    }
-
-    /// dot product, mostly used to compute residuals as sqrt(x*x)
-    Real operator*(const RigidCoord<2,real>& a) const
-    {
-        return center[0]*a.center[0]+center[1]*a.center[1]
-                +orientation*a.orientation;
-    }
-
-    /// Squared norm
-    Real norm2() const
-    {
-        Real angle = fmod( (Real) orientation, (Real) M_PI*2 );
-        return center*center + angle*angle;
-    }
-
-    /// Euclidean norm
-    real norm() const
-    {
-        return helper::rsqrt(norm2());
-    }
-
-    Vec2& getCenter () { return center; }
-    Real& getOrientation () { return orientation; }
-    const Vec2& getCenter () const { return center; }
-    const Real& getOrientation () const { return orientation; }
-
-    Vec2 rotate(const Vec2& v) const
-    {
-        Real s = sin(orientation);
-        Real c = cos(orientation);
-        return Vec2(c*v[0]-s*v[1],
-                s*v[0]+c*v[1]);
-    }
-    Vec2 inverseRotate(const Vec2& v) const
-    {
-        Real s = sin(-orientation);
-        Real c = cos(-orientation);
-        return Vec2(c*v[0]-s*v[1],
-                s*v[0]+c*v[1]);
-    }
-
-    Vec2 translate(const Vec2& v) const
-    {
-        return v + center;
-    }
-
-    static RigidCoord<2,real> identity()
-    {
-        RigidCoord<2,real> c;
-        return c;
-    }
-
-    /// Apply a transformation with respect to itself
-    void multRight( const RigidCoord<2,real>& c )
-    {
-        center += /*orientation.*/rotate(c.getCenter());
-        orientation = orientation + c.getOrientation();
-    }
-
-    /// compute the product with another frame on the right
-    RigidCoord<2,real> mult( const RigidCoord<2,real>& c ) const
-    {
-        RigidCoord<2,real> r;
-        r.center = center + /*orientation.*/rotate( c.center );
-        r.orientation = orientation + c.getOrientation();
-        return r;
-    }
-
-    template<class Mat>
-    void writeRotationMatrix( Mat& m) const
-    {
-        m[0][0] = (typename Mat::Real)cos(orientation); m[0][1] = (typename Mat::Real)-sin(orientation);
-        m[1][0] = (typename Mat::Real)sin(orientation); m[1][1] = (typename Mat::Real) cos(orientation);
-    }
-
-    /// Set from the given matrix
-    template<class Mat>
-    void fromMatrix(const Mat& m)
-    {
-        center[0] = m[0][2];
-        center[1] = m[1][2];
-        orientation = atan2(m[1][0],m[0][0]);
-    }
-
-    /// Write to the given matrix
-    template<class Mat>
-    void toMatrix( Mat& m) const
-    {
-        m.identity();
-        writeRotationMatrix( m );
-        m[0][2] = center[0];
-        m[1][2] = center[1];
-    }
-
-
-    /// create a homogeneous vector from a 2d vector
-    template <typename V>
-    static HomogeneousVec toHomogeneous( V v, real r=1. ){
-        return HomogeneousVec( v[0], v[1], r );
-    }
-    /// create a 2d vector from a homogeneous vector
-    template <typename V>
-    static Vec2 fromHomogeneous( V v ){
-        return Vec2( v[0], v[1] );
-    }
-
-
-    /// Write the OpenGL transformation matrix
-    void writeOpenGlMatrix( float m[16] ) const
-    {
-        //orientation.writeOpenGlMatrix(m);
-        m[0] = cos(orientation);
-        m[1] = sin(orientation);
-        m[2] = 0;
-        m[3] = 0;
-        m[4] = -sin(orientation);
-        m[5] = cos(orientation);
-        m[6] = 0;
-        m[7] = 0;
-        m[8] = 0;
-        m[9] = 0;
-        m[10] = 1;
-        m[11] = 0;
-        m[12] = (float)center[0];
-        m[13] = (float)center[1];
-        m[14] = (float)center[2];
-        m[15] = 1;
-    }
-
-    /// compute the projection of a vector from the parent frame to the child
-    Vec2 vectorToChild( const Vec2& v ) const
-    {
-        return /*orientation.*/inverseRotate(v);
-    }
-
-    /// Apply the transform to a point, i.e. project a point from the child frame to the parent frame (translation and rotation)
-    Vec2 projectPoint( const Vec2& v ) const
-    {
-        return rotate(v)+center;
-    }
-
-    /// Apply the transform to a vector, i.e. project a vector from the child frame to the parent frame (rotation only, no translation added)
-    Vec2 projectVector( const Vec2& v ) const
-    {
-        return rotate(v);
-    }
-
-    /// Apply the inverse transform to a point, i.e. project a point from the parent frame to the child frame (translation and rotation)
-    Vec2 unprojectPoint( const Vec2& v ) const
-    {
-        return inverseRotate(v-center);
-    }
-
-    ///  Apply the inverse transform to a vector, i.e. project a vector from the parent frame to the child frame (rotation only, no translation)
-    Vec2 unprojectVector( const Vec2& v ) const
-    {
-        return inverseRotate(v);
-    }
-
-    /// obsolete. Use projectPoint.
-    Vec2 pointToParent( const Vec2& v ) const { return projectPoint(v); }
-    /// obsolete. Use unprojectPoint.
-    Vec2 pointToChild( const Vec2& v ) const { return unprojectPoint(v); }
-
-
-    /// write to an output stream
-    inline friend std::ostream& operator << ( std::ostream& out, const RigidCoord<2,real>& v )
-    {
-        out<<v.center<<" "<<v.orientation;
-        return out;
-    }
-    /// read from an input stream
-    inline friend std::istream& operator >> ( std::istream& in, RigidCoord<2,real>& v )
-    {
-        in>>v.center>>v.orientation;
-        return in;
-    }
-    static Size max_size()
-    {
-        return 3;
-    }
-
-    /// Compile-time constant specifying the number of scalars within this vector (equivalent to the size() method)
-    static constexpr sofa::Size total_size = 3;
-
-    /// Compile-time constant specifying the number of dimensions of space (NOT equivalent to total_size for rigids)
-    static constexpr sofa::Size spatial_dimensions = 2;
-
-    real* ptr() { return center.ptr(); }
-    const real* ptr() const { return center.ptr(); }
-
-    static Size size() {return 3;}
-
-    /// Access to i-th element.
-    real& operator[](Size i)
-    {
-        if (i<2)
-            return this->center(i);
-        else
-            return this->orientation;
-    }
-
-    /// Const access to i-th element.
-    const real& operator[](Size i) const
-    {
-        if (i<2)
-            return this->center(i);
-        else
-            return this->orientation;
-    }
-
-    /// @name Tests operators
-    /// @{
-
-    bool operator==(const RigidCoord<2,real>& b) const
-    {
-        return center == b.center && orientation == b.orientation;
-    }
-
-    bool operator!=(const RigidCoord<2,real>& b) const
-    {
-        return center != b.center || orientation != b.orientation;
-    }
-
-    /// @}
-};
-
-template<class real>
-class RigidMass<2, real>
-{
-public:
-    typedef real value_type;
-    typedef real Real;
-    Real mass,volume;
-    Real inertiaMatrix;	      // Inertia matrix of the object
-    Real inertiaMassMatrix;    // Inertia matrix of the object * mass of the object
-    Real invInertiaMatrix;	  // inverse of inertiaMatrix
-    Real invInertiaMassMatrix; // inverse of inertiaMassMatrix
-    RigidMass(Real m=1)
-    {
-        mass = m;
-        volume = 1;
-        inertiaMatrix = 1;
-        recalc();
-    }
-    void operator=(Real m)
-    {
-        mass = m;
-        recalc();
-    }
-    void operator+=(Real m)
-    {
-        mass += m;
-        recalc();
-    }
-    void operator-=(Real m)
-    {
-        mass -= m;
-        recalc();
-    }
-    // operator to cast to const Real
-    operator const Real() const
-    {
-        return mass;
-    }
-    /// Mass for a circle
-    RigidMass(Real m, Real radius)
-    {
-        mass = m;
-        volume = radius*radius*R_PI;
-        inertiaMatrix = (radius*radius)/2;
-        recalc();
-    }
-    /// Mass for a rectangle
-    RigidMass(Real m, Real xwidth, Real ywidth)
-    {
-        mass = m;
-        volume = xwidth*xwidth + ywidth*ywidth;
-        inertiaMatrix = volume/12;
-        recalc();
-    }
-
-    void recalc()
-    {
-        inertiaMassMatrix = inertiaMatrix * mass;
-        if (inertiaMatrix == 0.)
-        {
-            throw std::runtime_error("Attempt to divide by zero");
-        }
-        invInertiaMatrix = 1. / inertiaMatrix;
-        if (inertiaMassMatrix == 0.)
-        {
-            throw std::runtime_error("Attempt to divide by zero");
-        }
-        invInertiaMassMatrix = 1. / inertiaMassMatrix;
-    }
-    inline friend std::ostream& operator << (std::ostream& out, const RigidMass<2,Real>& m )
-    {
-        out<<m.mass;
-        out<<" "<<m.volume;
-        out<<" "<<m.inertiaMatrix;
-        return out;
-    }
-    inline friend std::istream& operator >> (std::istream& in, RigidMass<2,Real>& m )
-    {
-        in>>m.mass;
-        in>>m.volume;
-        in>>m.inertiaMatrix;
-        return in;
-    }
-    void operator *=(Real fact)
-    {
-        mass *= fact;
-        inertiaMassMatrix *= fact;
-        invInertiaMassMatrix /= fact;
-    }
-    void operator /=(Real fact)
-    {
-        mass /= fact;
-        inertiaMassMatrix /= fact;
-        invInertiaMassMatrix *= fact;
-    }
-};
-
-template<typename real>
-inline RigidDeriv<2,real> operator*(const RigidDeriv<2,real>& d, const RigidMass<2,real>& m)
-{
-    RigidDeriv<2,real> res;
-    getVCenter(res) = getVCenter(d) * m.mass;
-    getVOrientation(res) = m.inertiaMassMatrix * getVOrientation(d);
-    return res;
-}
-
-template<typename real>
-inline RigidDeriv<2,real> operator/(const RigidDeriv<2,real>& d, const RigidMass<2, real>& m)
-{
-    RigidDeriv<2,real> res;
-    getVCenter(res) = getVCenter(d) / m.mass;
-    getVOrientation(res) = m.invInertiaMassMatrix * getVOrientation(d);
-    return res;
-}
-
-
 /** Degrees of freedom of 2D rigid bodies.
 */
 template<typename real>
@@ -1723,6 +264,7 @@ public:
     typedef RigidDeriv<2,Real> Deriv;
     typedef RigidCoord<2,Real> Coord;
     typedef Real AngularVector;
+    typedef MapMapSparseMatrix<Deriv> MatrixDeriv;
 
     static constexpr sofa::Size spatial_dimensions = Coord::spatial_dimensions;
     static constexpr sofa::Size coord_total_size = Coord::total_size;
@@ -1747,8 +289,6 @@ public:
     typedef type::vector<Coord> VecCoord;
     typedef type::vector<Deriv> VecDeriv;
     typedef type::vector<Real> VecReal;
-
-    typedef MapMapSparseMatrix<Deriv> MatrixDeriv;
 
     template<typename T>
     static void set(Coord& c, T x, T y, T)
@@ -1858,8 +398,6 @@ public:
 
 };
 
-
-
 template<> constexpr const char* Rigid2dTypes::Name() { return "Rigid2d"; }
 template<> constexpr const char* Rigid2fTypes::Name() { return "Rigid2f"; }
 
@@ -1910,10 +448,4 @@ static void rigidTransform ( V1& points, V2& velocities, SReal tx, SReal ty, SRe
 }
 //@}
 
-
-}
-
-} // namespace sofa
-
-
-#endif
+} // namespace sofa::defaulttype
